requirements.txt
asgiref             3.5.2
Django              4.1.1
djangorestframework 3.14.0
mysqlclient         2.1.1
pip                 22.3.1
PyMySQL             1.0.2
pytz                2022.5
setuptools          63.2.0
sqlparse            0.4.3
tzdata              2022.5
django-filter       22.1    
markdown            3.4.1
dj-rest-auth        2.2.7